pyproject.toml
[tool.poetry]
name = "invoice-app"
version = "0.1.0"
description = ""
authors = ["RashminDungrani <rashmin8248@gmail.com>"]
readme = "README.md"
packages = [{include = "invoice_app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.87.0"
sqlmodel = "^0.0.8"
aiomysql = "^0.1.1"
uvicorn = "^0.19.0"
pytest = "^7.2.0"
httpx = "^0.23.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.5"
yarl = "^1.8.1"
alembic = "^1.8.1"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
